flask/requirements.txt
redis==4.3.5
Flask==2.2.2
orjson==3.8.0
pydantic[dotenv]==1.10.2
SQLAlchemy==1.4.44
psycopg2-binary==2.9.5
gevent==22.10.2
uWSGI==2.0.21
pycryptodomex==3.15.0
PyJWT==2.6.0
psycogreen==1.0.2
alembic==1.8.1
requests==2.28.1
opentelemetry-api==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-instrumentation-flask==0.35b0
opentelemetry-exporter-jaeger==1.14.0
